// src/util/types.ts
export interface GuildRecordSettings {
    // bot access
    allowedChannelsType: 'blacklist' | 'whitelist',
    allowedChannelsWhitelist: string[],
    allowedChannelsBlacklist: string[],
    allowedRolesType: 'blacklist' | 'whitelist',
    allowedRolesWhitelist: string[],
    allowedRolesBlacklist: string[],
    blockedUsers: string[],

    // role management
    persistentRoles: string[],
    timedRoles: Record<string, number>,

    // greetings
    greetingsEnabled: boolean,
    greetingsChannel: string|null,
    greetingsMessages: {
        join: string,
        leave: string,
    },

    // moderation
    moderationEnabled: boolean,
    moderationPenaltyLevels: ModerationPenaltyLevel[],
    moderationLogChannel: string|null,
    moderatorRoles: string[],

    // music
    musicEnabled: boolean,
    musicDjRole: string|null,
    musicEveryoneDj: boolean,
    musicSkipVotesRequired: number,
    
    // flags
    perCommandSettings: Record<string, BotCommandSettings>
}
export interface ModerationPenaltyLevel {
    type: 'verbalWarn' | 'warn' | 'mute' | 'kick' | 'ban',
    duration: number,
    warnRole: string,
    muteOrBanDuration: number,
}
export interface BotCommandSettings {
    disabled: boolean
    allowedChannelsType: 'blacklist' | 'whitelist'
    allowedChannelsWhitelist: string[]
    allowedChannelsBlacklist: string[]
    allowedRolesType: 'blacklist' | 'whitelist'
    allowedRolesWhitelist: string[]
    allowedRolesBlacklist: string[]
    blockedUsers: string[]
}